requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
# backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==40.0.2
defusedxml==0.7.1
dj-rest-auth==4.0.0
django-allauth==0.52.0
django-countries==7.5.1
django-phonenumber-field==7.1.0
django-randompinfield==0.4.7
djangorestframework-simplejwt==5.2.2
frozenlist==1.3.3
idna==3.4
kombu==5.2.4
multidict==6.0.4
oauthlib==3.2.2
phonenumbers==8.13.11
Pillow==9.5.0
prompt-toolkit==3.0.38
pycparser==2.21
PyJWT==2.7.0
python3-openid==3.2.0
pytz==2023.3
redis==4.5.5
requests==2.30.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
twilio==8.2.1
typing_extensions==4.6.2
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2
